* Some 64-bit patches * Better handling of uncaught exceptions * Print singletons using printMessage()

+++ src/pl-file.c
@@ -292,7 +292,7 @@ initIO()
   getStreamContext(Soutput);
   getStreamContext(Serror);
   for( i=0, np = standardStreams; *np; np++, i++ )
-    addHTable(streamAliases, (void *)*np, (void *)i);
+    addHTable(streamAliases, (void *)*np, (void *)(long)i);
 
   GD->io_initialised = TRUE;
 }
